Add files via upload
This is the version I've got so far. It contains the red cubes(plants) and blue capsules(herbivores). I've left a folder containing the old scripts from pre-modularization; using them is not recommended. Ive also added the changelog, which is more of a to-do list. It contains things I want to add, things I want to add if I'm really bored, and things I've already added.